Fix heredoc/nowdoc in multiline arrays.
The heredoc/nowdoc closing identifier must be the only content
on the line. The sniff was incorrectly adding the comma
directly after the closing identifier of the last element,
introducing a syntax error.
The sniff also expected the comma to be directly after
the closing identifier for every other element.

The comma must be on the next line, and it should
be indented to the same depth as array indices.